Nest tested resource in namespace
