+++ test_apps/rails_51/config/routes.rb
@@ -1,6 +1,8 @@
 Rails.application.routes.draw do
-  resources :martians do
-    resources :pets
+  namespace :v1 do
+    resources :martians do
+      resources :pets
+    end
   end
   # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
 end
