+++ test_apps/rails_51/spec/murker/v1/martians/__id/GET.yml
@@ -1,7 +1,7 @@
 ---
 openapi: 3.0.0
 paths:
-  "/martians/{id}":
+  "/v1/martians/{id}":
     get:
       responses:
         "'200'":
