+++ test_apps/rails_51/spec/requests/martians_spec.rb
@@ -1,10 +1,15 @@
 require 'rails_helper'
+require 'murker/spec_helper'
 
-RSpec.describe "Martians", type: :request do
-  describe "GET /martians" do
-    it "works! (now write some real specs)" do
-      get martians_path
-      expect(response).to have_http_status(200)
+RSpec.describe V1::MartiansController, type: :request do
+
+  describe "GET #index and martian" do
+    it "returns a success response", :murker do
+      martian = Martian.create! name: 'spajic', age: 30, id: 1
+
+      get '/v1/martians.json'
+
+      expect(response).to be_success
     end
   end
 end
