--- test_apps/rails_51/spec/views/martians/show.html.erb_spec.rb
@@ -1,16 +1,0 @@
-require 'rails_helper'
-
-RSpec.describe "martians/show", type: :view do
-  before(:each) do
-    @martian = assign(:martian, Martian.create!(
-      :name => "Name",
-      :age => 2
-    ))
-  end
-
-  it "renders attributes in <p>" do
-    render
-    expect(rendered).to match(/Name/)
-    expect(rendered).to match(/2/)
-  end
-end
